# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio centralizado de autenticación y autorización con JWT RS256, RBAC y auditoría.",
        "title": "Sentinel API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/authz/me/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos los roles, permisos y centros de costo del usuario autenticado para la aplicación actual.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autorización"
                ],
                "summary": "Mis permisos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permisos del usuario",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerMePermissionsResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/authz/permissions-map": {
            "get": {
                "security": [
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Retorna el mapa completo de permisos de la aplicación, firmado con RSA-SHA256.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autorización"
                ],
                "summary": "Mapa de permisos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mapa de permisos firmado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Aplicación no encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/authz/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "AppKeyAuth": []
                    }
                ],
                "description": "Verifica si el usuario autenticado tiene un permiso específico, opcionalmente en un centro de costo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autorización"
                ],
                "summary": "Verificar permiso",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave secreta de la aplicación",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token JWT. Formato: Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permiso a verificar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerVerifyPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de la verificación",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerVerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.SwaggerErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {},
                "message": {
                    "type": "string",
                    "example": "campo requerido"
                }
            }
        },
        "internal_handler.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_handler.SwaggerErrorDetail"
                }
            }
        },
...
        "internal_handler.SwaggerMePermissionsResponse": {
            "type": "object",
            "properties": {
                "application": {
                    "type": "string",
                    "example": "sentinel"
                },
                "cost_centers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "temporary_roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "internal_handler.SwaggerVerifyPermissionRequest": {
            "type": "object",
            "properties": {
                "cost_center_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "permission": {
                    "type": "string",
                    "example": "admin.users.read"
                }
            }
        },
        "internal_handler.SwaggerVerifyResponse": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean",
                    "example": true
                },
                "evaluated_at": {
                    "type": "string",
                    "example": "2025-01-15T10:30:00Z"
                },
                "permission": {
                    "type": "string",
                    "example": "admin.users.read"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        }
    },
    "securityDefinitions": {
        "AppKeyAuth": {
            "description": "Clave secreta de la aplicación cliente",
            "type": "apiKey",
            "name": "X-App-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Token JWT de acceso. Formato: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}